Update : DietAction 유효성검사

+++ src/main/java/diet/controller/action/DeleteDietAction.java
@@ -12,29 +12,61 @@ import diet.model.DietDao;
 import user.controller.Action;
 
 public class DeleteDietAction implements Action{
-	
+
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-		int dietIndex = Integer.parseInt(request.getParameter("dietIndex"));
-		DietDao dao = DietDao.getInstance();
-		
 		response.setContentType("application/json");
 		response.setCharacterEncoding("UTF-8");
-		
-		JSONObject object = new JSONObject();
-		
+
 		try {
+			int dietIndex = Integer.parseInt(request.getParameter("dietIndex"));
+
+			validateInput(dietIndex);
+
+			DietDao dao = DietDao.getInstance();
+
+			if (!dao.existsById(dietIndex)) {
+				throw new ValidationException("Diet with the given index does not exist");
+			}
+
 			dao.deleteDiet(dietIndex);
-			object.put("status", 200);
-			object.put("message", "식단 삭제 완료");
+
+			JSONObject jsonResponse = new JSONObject();
+			jsonResponse.put("status", 200);
+			jsonResponse.put("message", "식단 삭제 완료");
+			response.getWriter().write(jsonResponse.toString());
+
 			System.out.println("식단 삭제 완료");
+
+		} catch (ValidationException e) {
+			JSONObject jsonResponse = new JSONObject();
+			jsonResponse.put("status", 400);
+			jsonResponse.put("message", e.getMessage());
+			response.getWriter().write(jsonResponse.toString());
+
+			System.out.println("유효성 검사 실패: " + e.getMessage());
+
 		} catch (Exception e) {
 			e.printStackTrace();
-			object.put("status", 500);
-			object.put("message", "식단 삭제 실패");
+			JSONObject jsonResponse = new JSONObject();
+			jsonResponse.put("status", 500);
+			jsonResponse.put("message", "식단 삭제 실패");
+			response.getWriter().write(jsonResponse.toString());
+
 			System.out.println("식단 삭제 실패");
 		}
-		response.getWriter().write(object.toString());
+	}
+
+	private void validateInput(int dietIndex) throws ValidationException {
+		if (dietIndex <= 0) {
+			throw new ValidationException("Invalid diet index");
+		}
+	}
+
+	// ValidationException 클래스 정의
+	public class ValidationException extends Exception {
+		public ValidationException(String message) {
+			super(message);
+		}
 	}
 }
